new user registration form almost working
