+++ src/App.js
@@ -23,10 +23,36 @@ function App() {
   const url = "https://hustle-busters.herokuapp.com"
   // const url = "http://localhost:4000"
 
+  //STATES
   const [leads, setLeads] = React.useState([])
   const [divisions, setDivisions] = React.useState([])
   const [searchDivision, setSearchDivision] = React.useState("")
+  //new user registration form
+  const [regData, setRegData] = React.useState()
 
+  //EMPTY USER
+  const emptyUser = {
+      firstName: "",
+      lastName: "",
+      email: "",
+      userName: "",
+      password: ""
+  }
+
+  //CREATE USER
+  const createUser = (newUser) => {
+    fetch(url + "/usernames/", {
+        method: "post",
+        headers: {
+            "Content-Type" : "application/json"
+        },
+        body: JSON.stringify(newUser)
+    })
+    }
+
+  
+  
+  //GET LEADS
   const getLeads = () => {
     fetch(url + "/leads")
     .then(response => response.json())
@@ -39,7 +65,7 @@ function App() {
     getLeads()
   }, [])
 
-
+  //GET DIVISIONS
 const getDivisions = () => {
   fetch(url + '/divisions')
   .then((response) => response.json())
@@ -49,6 +75,7 @@ const getDivisions = () => {
   })
 }
 
+//GET DIVISIONS BY NAME
 const getDivisionByName = (searchDivision) => {
   console.log('searched division', searchDivision)
   fetch(url + '/divisions/name/' + searchDivision)
@@ -82,7 +109,7 @@ React.useEffect(() => {
         </Route>
         <Route
           path='/user-registration'
-          render={(rp) => <UserReg />}>
+          render={(rp) => <UserReg {...rp} user={emptyUser} handleSubmit={createUser} />}>
         </Route>
         <Route
           path='/division-leads'
